Fix formula for calculating delta
Currently, as I understand, the formula takes July 31st Base (B1) and Current Membership as of July 31st (C1), September 30th base (B2) and September 30th Current Membership (C2) as follows:

If (C2-B1)-(C1-B1) >= 7, then club is a Smedley Stretch winner

It seems that when members are "reinstated" with lerss than a 6th month lapse, then it is not considered a reinstatement, but just a late renewal. Thus the formula needs to be:

If (C2-B2)-(C1-B1) >=7, then club is a Smedley Stretch winner.