Handle loading historical clubs if only clubs was reset
If the CLUBS table was the only one reset (not the performance tables),
we need to check each day's update with the previous day's status, not
the status for the latest day available in the performance tables.

This should ONLY matter when adding a new field to the CLUBS table and
catching up.