Change all program parameters to lowercase
A few program parameters were formerly in camelCase to make them
slightly easier to read.  However, with the change to use an INI
file instead of a YAML file for most programs, parameter names
have to be all lower-case to match what comes from the INI file,
so I changed them all.  And in the few cases where I was using a
parameter name (newAlignment, for example) as a keyword parameter
to an internal routine, I changed it to lowercase for consistency.

@@TODO@@ Change awardmail to use the INI file instead of its own
YAML file.